add api authentication with sanctum

+++ app/Http/Controllers/Api/ProductsController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\ProductResource;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Response;
 
 class ProductsController extends Controller
@@ -13,6 +14,13 @@ class ProductsController extends Controller
     /**
      * Display a listing of the resource.
      */
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')
+                ->except('index','show');
+    }
+
     public function index(Request $request)
     {
          $products = Product::with(['category:id,name','store:id,name','tags:id,name'])
@@ -35,6 +43,12 @@ class ProductsController extends Controller
             'compare_price' => 'nullable|numeric|gt:price',
         ]);
 
+        $user=$request->user();
+        if (!$user->tokenCan('products.create')){
+            return \response([
+                'message'=>'Not allowed'
+            ],403);
+        }
         $product = Product::create($request->all());
 
         return Response::json($product, 201, [
@@ -65,7 +79,12 @@ class ProductsController extends Controller
             'compare_price' => 'nullable|numeric|gt:price',
         ]);
 
-
+        $user=$request->user();
+        if (!$user->tokenCan('products.update')){
+            return \response([
+                'message'=>'Not allowed'
+            ],403);
+        }
         $product->update($request->all());
 
 
@@ -77,6 +96,13 @@ class ProductsController extends Controller
      */
     public function destroy(string $id)
     {
+        $user=Auth::guard('sanctum')->user();
+        if (!$user->tokenCan('products.delete')){
+              return \response([
+                  'message'=>'Not allowed'
+              ],403);
+        }
+
         Product::destroy($id);
         return [
             'message' => 'Product deleted successfully',
